+++ re_contact_book_system/backend/migrations/versions/7e216e07dc5e_initial_tables.py
@@ -31,7 +31,7 @@ def upgrade():
         mysql_collate="utf8mb4_unicode_ci",
     )
 
-    # grades
+    # grades(学年)
     op.create_table(
         "grades",
         sa.Column("id", sa.Integer, primary_key=True),